ppc64: Fix <sys/user.h> include order
The version of the PPC64 sys/user.h header on CentOS 7.7 is not
self-contained: it uses the size_t type but doesn't pull in any
other header that defines it. This results in a compile error:


In file included from risu_ppc64.c:14:
/usr/include/sys/user.h:27:9: error: unknown type name ‘size_t’
    27 |         size_t          u_tsize;                /* text size (pages) */
       |         ^~~~~~

Work around this by including sys/user.h only after risu.h (which
will bring in the system header defining size_t).

Signed-off-by: Richard Henderson <richard.henderson@linaro.org>
Reviewed-by: Philippe Mathieu-Daudé <philmd@linaro.org>
Message-id: 20240511115400.7587-2-richard.henderson@linaro.org
[PMM: update commit message]
Signed-off-by: Peter Maydell <peter.maydell@linaro.org>